+++ src/rules/index.js
@@ -1,0 +1,7 @@
+// Adapted with gratitude from eslint-plugin-filenames (https://github.com/selaux/eslint-plugin-filenames) and used under an MIT license
+
+module.exports = {
+  rules: {
+    'match-regex': require('./match-regex'),
+  },
+}